Move devsetup agent and tests to api_lib
